+++ src/components/ColorPicker/index.tsx
@@ -5,6 +5,7 @@ import React, { useState } from 'react';
 import { ColorPickerModal } from './modal';
 
 import { Color, Colors } from 'bibata/app';
+import { GemsSVG } from '@components/svgs';
 
 type Props = {
   name: string;
@@ -12,46 +13,60 @@ type Props = {
   selected?: boolean;
   disabled?: boolean;
   onClick?: () => void;
-  children?: React.ReactNode;
 };
 
 export const ColorPickerButton: React.FC<Props> = (props) => {
+  const { name, color, selected, disabled } = props;
+
   const stops = Array.from(
     { length: 12 },
     (_, i) => `hsl(${i * (360 / 12)}, 100%, 50%)`
   ).join(', ');
+  const wheel = `radial-gradient( circle closest-side, hsl(0, 0%, 100%), hsl(0, 0%, 0%) 90%), conic-gradient(${stops})`;
 
   return (
     <button
-      className={`container p-4 sm:p-6 flex flex-col gap-1 justify-center items-center rounded-2xl sm:rounded-3xl ring-1 ${
-        props.selected ? 'ring-white/[.3] bg-white/[.1]' : 'ring-white/[.2]'
-      }`}
-      title={
-        props.name !== 'Custom'
-          ? `Bibata ${props.name}`
-          : 'Customize Bibata Colors'
+      className={`p-1.5 flex flex-col gap-1 justify-center items-center rounded-xl sm:rounded-2xl transform ${
+        selected ? '-translate-y-4 scale-100 font-bold' : 'scale-95'
+      } transition-all ease-in-out duration-100`}
+      title={name !== 'Custom' ? `Bibata ${name}` : 'Customize Bibata Colors'}
+      style={
+        selected
+          ? color
+            ? {
+                backgroundColor: color.base,
+                color: color.outline,
+                boxShadow: `0 0 12px 0 ${color.base}77`
+              }
+            : {
+                backgroundColor: '#010101',
+                color: '#ffffff',
+                boxShadow:
+                  '0 4px 6px rgba(0, 0, 0, 0.1), 0 6px 15px rgba(0, 0, 0, 0.1), 0 9px 24px rgba(255, 0, 0, 0.5), 0 12px 30px rgba(0, 255, 0, 0.5), 0 15px 36px rgba(0, 0, 255, 0.5)'
+              }
+          : {}
       }
-      disabled={props.selected && props.disabled}
+      disabled={selected && disabled}
       onClick={props.onClick}>
       <div
-        className='w-10 sm:w-20 h-10 sm:h-20 rounded-full flex justify-center items-center'
+        className={`w-full h-24 md:h-32 flex justify-center items-center ${
+          selected ? 'rounded-lg sm:rounded-xl' : 'rounded-2xl sm:rounded-3xl'
+        }`}
         style={
-          props.color
+          color
             ? {
-                color: props.color.outline,
-                backgroundColor: props.color.base
+                color: color.outline,
+                backgroundColor: color.base
               }
             : {
                 color: 'transparent',
                 backgroundBlendMode: 'screen',
-                background: `radial-gradient( circle closest-side, hsl(0, 0%, 100%), hsl(0, 0%, 0%) 90%), conic-gradient(${stops})`
+                background: wheel
               }
         }>
-        {props.children}
+        <GemsSVG gems={name} />
       </div>
-      <div className='mt-0 sm:mt-3 text-center text-xs sm:text-sm'>
-        {props.name}
-      </div>
+      <div className='mt-0 sm:mt-3 text-center text-xs sm:text-sm'>{name}</div>
     </button>
   );
 };
@@ -70,7 +85,7 @@ export const ColorPicker: React.FC<ColorPickerProps> = (props) => {
   return (
     <>
       <div className='flex items-center justify-center'>
-        <div className='w-full md:w-2/3 lg:w-1/2 sm:mx-32 grid grid-cols-4 gap-3 sm:gap-7'>
+        <div className='w-full md:w-2/3 lg:w-1/2 sm:mx-32 grid grid-cols-4 gap-2 sm:gap-7'>
           {Object.entries(props.colors).map(([name, color], i) => (
             <ColorPickerButton
               key={i}