+++ src/components/ColorPicker/modal.tsx
@@ -1,7 +1,10 @@
-import React, { useState } from 'react';
+'use client';
+
+import React, { useEffect, useState } from 'react';
 import Wheel from '@uiw/react-color-wheel';
 
 import { CursorPreview } from './preview';
+import { CloseSVG, RefreshSVG } from '@components/svgs';
 
 import { generateRandomColors } from '@utils/randomColors';
 
@@ -24,13 +27,13 @@ const ColorWheelCard: React.FC<ColorWheelCardProps> = (props) => {
 
   return (
     <div className='flex flex-col justify-center items-center'>
-      <div className='font-bold'>{props.title}</div>
+      <div className='font-bold text-xs sm:text-md'>{props.title}</div>
       <div className='text-xl sm:text mt-2 p-2'>{props.children}</div>
       <input
         type='text'
         minLength={1}
         maxLength={7}
-        className='w-3/4 sm:w-full mt-2 p-1 bg-white/[.05] text-center border border-white/[.1] hover:border-white focus:outline-none rounded-2xl'
+        className='w-4/5 sm:w-full text-xs sm:text-md mt-2 p-1 bg-white/[.05] text-center border border-white/[.1] hover:border-white focus:outline-none rounded-2xl'
         value={props.value}
         placeholder='e.g., #ff0000'
         onChange={(e) => {
@@ -57,6 +60,14 @@ type ColorPickerModalProps = {
 };
 
 export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
+  const [wheelSize, setWheelSize] = useState(() => {
+    try {
+      return window.innerWidth <= 500 ? 65 : 90;
+    } catch {
+      return 90;
+    }
+  });
+
   const defaultColor = generateRandomColors();
   const [baseColor, setBaseColor] = useState<string>(defaultColor[0]);
   const [outlineColor, setOutlineColor] = useState<string>(defaultColor[1]);
@@ -79,42 +90,54 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
     props.onClose();
   };
 
+  useEffect(() => {
+    if (window !== undefined) {
+      const handleResize = () => {
+        setWheelSize(window.innerWidth <= 500 ? 65 : 90);
+      };
+
+      window.addEventListener('resize', handleResize);
+
+      return () => {
+        window.removeEventListener('resize', handleResize);
+      };
+    }
+  }, []);
+
   return (
     <>
       {props.isOpen && (
         <div
           onClick={(e) => e.target === e.currentTarget && props.onClose()}
-          className='z-20 fixed top-0 left-0 w-full h-full flex items-center justify-center bg-black bg-opacity-80'>
-          <div className='bg-[#333333] w-full md:w-1/2 xl:w-1/3 max-h-full overflow-y-auto p-4 m-4 rounded-3xl shadow-lg'>
+          className='z-20 fixed top-0 left-0 w-full h-full flex items-center justify-center bg-black backdrop-filter backdrop-blur-xl firefox:bg-opacity-80 bg-opacity-80'>
+          <div className='w-full md:w-1/2 xl:w-1/3 max-h-full overflow-y-auto p-4 m-4 rounded-3xl shadow-lg bg-black ring-1 ring-white/[.06]'>
             <div className='flex justify-between text-xs'>
-              <div>
-                <button
-                  className='p-2 bg-[#93f5d2] text-black rounded-2xl font-bold hover:bg-[#d9c57f] active:bg-[#ffc68f]'
-                  onClick={refreshColors}>
-                  Magic
-                </button>
-              </div>
               <button
-                className='p-2 bg-gray-500 text-white rounded-full font-bold hover:bg-gray-400'
+                className='p-2 bg-[#93f5d2] text-black rounded-2xl hover:bg-[#d9c57f] active:bg-[#ffc68f]'
+                onClick={refreshColors}>
+                <RefreshSVG />
+              </button>
+              <button
+                className='p-3 bg-gray-500 text-white rounded-2xl font-bold hover:bg-gray-400'
                 onClick={props.onClose}>
-                X
+                <CloseSVG />
               </button>
             </div>
-            <div className='h-96 md:h-72 mt-2'>
+            <div className='h-48 sm:h-96 md:h-72 mt-2'>
               <CursorPreview
                 base={baseColor}
                 outline={outlineColor}
                 watch={watchColor}
               />
             </div>
-            <div className='mt-8 grid sm:grid-cols-3 gap-10'>
+            <div className='mt-8 grid grid-cols-3 gap-10'>
               <ColorWheelCard
                 title='Base'
                 value={baseColor}
                 onChange={(c) => setBaseColor(c)}>
                 <Wheel
-                  width={100}
-                  height={100}
+                  width={wheelSize}
+                  height={wheelSize}
                   color={baseColor}
                   onChange={(cr) => {
                     setBaseColor(cr.hex);
@@ -127,8 +150,8 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
                 value={outlineColor}
                 onChange={(c) => setOutlineColor(c)}>
                 <Wheel
-                  width={100}
-                  height={100}
+                  width={wheelSize}
+                  height={wheelSize}
                   color={outlineColor}
                   onChange={(cr) => {
                     setOutlineColor(cr.hex);
@@ -141,8 +164,8 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
                 value={watchColor}
                 onChange={(c) => setWatchColor(c)}>
                 <Wheel
-                  width={100}
-                  height={100}
+                  width={wheelSize}
+                  height={wheelSize}
                   color={watchColor}
                   onChange={(cr) => {
                     setWatchColor(cr.hex);
@@ -152,7 +175,7 @@ export const ColorPickerModal: React.FC<ColorPickerModalProps> = (props) => {
             </div>
             <div className='mt-11 flex justify-center'>
               <button
-                className='w-1/2 py-4 bg-green-600 text-white font-bold rounded-3xl text-xl hover:bg-green-500'
+                className='w-36 py-3 bg-green-600 text-white font-bold rounded-2xl text-sm sm:text-md hover:bg-green-500'
                 onClick={handleColorPick}>
                 Apply
               </button>