+++ src/components/Cursors/loading.tsx
@@ -16,9 +16,9 @@ export const CursorsLoading: React.FC<Props> = (_props) => {
             className='w-full mb-4 overflow-hidden rounded-2xl sm:rounded-3xl bg-white/[0.05] border-white/[.1] border'>
             <div
               className={
-                'w-full h-20 sm:h-40 animate-pulse bg-white/[.2]'
+                'w-full h-24 sm:h-40 animate-pulse bg-white/[.2]'
               }></div>
-            <div className='flex animate-pulse bg-white/[.1] h-8'></div>
+            <div className='flex animate-pulse bg-white/[.1] h-10' />
           </div>
         ))}
       </div>