+++ src/components/NavBar/index.tsx
@@ -17,30 +17,27 @@ export const NavBar: React.FC<Props> = (_props) => {
   const { data: session, status } = useSession();
 
   return (
-    <header className='bg-transparent p-2 sm:p-5 top-0 w-full h-15'>
-      <div className='flex items-center justify-between'>
-        <div className='overflow-hidden flex items-center justify-center gap-4'>
-          <Link
-            href='/'
-            title='Goto Homepage'
-            className='bg-blue-800 p-2 rounded-xl sm:rounded-3xl'>
+    <header className='sticky py-px top-0 z-20 bg-[#030303] backdrop-filter backdrop-blur-2xl border-b border-white/[.1] firefox:bg-opacity-70 bg-opacity-70'>
+      <nav className='container mx-auto p-3 md:p-4 flex items-center justify-between'>
+        <Link href='/'>
+          <div className='overflow-hidden flex items-center justify-center gap-2'>
             <BibataLogo />
-          </Link>
-          <span className='inline-flex items-center gap-1'>
-            {pathname === '/studio' && (
-              <span className='text-lg sm:text-3xl font-bold'>Studio</span>
-            )}
-            {session?.user?.role === 'PRO' && <ProBadge size={18} />}
-          </span>
-        </div>
+            <span className='inline-flex items-center gap-1'>
+              {pathname === '/studio' && (
+                <span className='text-2xl font-bold'>Studio</span>
+              )}
+              {session?.user?.role === 'PRO' && <ProBadge size={18} />}
+            </span>
+          </div>
+        </Link>
         {status === 'loading' ? (
-          <div className='m-2 w-24 mr-4'>
-            <div className='rounded-full h-8 ring-white/[.2] ring-1 p-1 animate-pulse bg-white/[.1]' />
+          <div className='m-1 sm:m-1 w-16 sm:w-24 mr-1 sm:mr-4'>
+            <div className='rounded-full h-6 sm:h-8 ring-white/[.2] ring-1 p-1 animate-pulse bg-white/[.1]' />
           </div>
         ) : (
           <Profile session={session} />
         )}
-      </div>
+      </nav>
     </header>
   );
 };