+++ src/components/NavBar/profile.tsx
@@ -17,12 +17,12 @@ export const Profile: React.FC<Props> = (props) => {
     <div className='inline-flex items-center gap-1'>
       {!props.session ? (
         <button
-          className='inline-flex items-center py-1 sm:py-2 px-3 sm:px-6 bg-white/[.03] ring-1 ring-white/[.2] shadow text-white fill-white rounded-xl sm:rounded-2xl text-lg font-semibold'
+          className='inline-flex items-center justify-center py/[3px] sm:py-1 px-3 sm:px-4 bg-white/[.03] ring-1 ring-white/[.2] shadow text-white fill-white rounded-xl text-lg font-semibold'
           title='Click to Connect your GitHub Account'
           onClick={() => signIn('github', { callbackUrl: '/' })}>
-          <span className='mr-2 block sm:hidden text-xl font-bold'>+</span>
+          <span className='mr-2 block sm:hidden text-sm font-bold'>+</span>
           <GitHubLogo />
-          <span className='ml-3 hidden sm:block'>Connect</span>
+          <span className='ml-3 hidden sm:block text-sm'>Connect</span>
         </button>
       ) : (
         <>
@@ -36,7 +36,7 @@ export const Profile: React.FC<Props> = (props) => {
             <LogoutSVG />
           </button>
 
-          <div className='w-7 sm:w-14 h-7 sm:h-14 overflow-hidden rounded-3xl ring-white/[.2] ring-1 '>
+          <div className='w-9 sm:w-12 overflow-hidden rounded-2xl ring-white/[.2] ring-1 '>
             <img
               width={100}
               height={100}