+++ src/components/Tooltip.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 
 type Props = {
-  content: string;
+  tooltip: string;
   children: React.ReactElement;
 };
 
@@ -18,8 +18,13 @@ const Tooltip: React.FC<Props> = (props) => {
       onMouseLeave={() => setShowTooltip(false)}
       onClick={() => setShowTooltip(!showTooltip)}>
       {showTooltip && (
-        <div className='absolute w-96 top-1/2 left-1/2 transform -translate-x-1/2 translate-y-4 py-2 px-4 bg-black text-white rounded-xl shadow'>
-          {props.content}
+        <div className='absolute w-52 sm:w-64 top-1/2 left-1/2  transform -translate-x-1/2 translate-y-4  shadow'>
+          <div className='flex max-w-xs flex-col items-center shadow-lg'>
+            <div className='clip-bottom h-2 w-4 bg-black' />
+            <div className='rounded-xl bg-black p-2 text-[8px] sm:text-xs text-white'>
+              {props.tooltip}
+            </div>
+          </div>
         </div>
       )}
       {props.children}