+++ src/components/svgs/index.ts
@@ -12,6 +12,14 @@ import { LockSVG } from './lock';
 import { LinuxDownloadSVG } from './linux';
 import { WindowsDownloadSVG } from './windows';
 
+import { RefreshSVG } from './refresh';
+import { CloseSVG } from './close';
+
+import { ModernSVG } from './modern';
+import { OriginalSVG } from './original';
+
+import { GemsSVG } from './gems';
+
 export {
   BibataLogo,
   GitHubLogo,
@@ -23,5 +31,10 @@ export {
   ErrorSVG,
   InfoSVG,
   WindowsDownloadSVG,
-  LinuxDownloadSVG
+  LinuxDownloadSVG,
+  RefreshSVG,
+  CloseSVG,
+  ModernSVG,
+  OriginalSVG,
+  GemsSVG
 };