+++ src/components/svgs/windows.tsx
@@ -8,9 +8,7 @@ type Props = {};
 export const WindowsDownloadSVG: React.FC<Props> = (_props) => {
   return (
     <svg
-      width='40'
-      height='40'
-      className='fill-inherit'
+      className='fill-current w-16 sm:w-24'
       viewBox='0 0 75 75'
       fill='none'
       xmlns='http://www.w3.org/2000/svg'>