+++ src/types/bibata.d.ts
@@ -31,6 +31,11 @@ declare module 'bibata/app' {
     frames: string[];
     delay: number;
   };
+
+  type ErrorLogs = {
+    text: string;
+    [k: string]: any;
+  };
 }
 
 declare module 'bibata/misc' {
@@ -89,7 +94,6 @@ declare module 'bibata/core-api/responses' {
   };
 
   type UploadResponse = {
-    status: number;
     id: string;
     files: string[];
     error: string[];
@@ -105,8 +109,12 @@ declare module 'bibata/core-api/responses' {
   };
 
   type DownloadError = {
-    status: number;
     id: string;
     error: string[];
+  };
+
+  type DownloadFile = {
+    blob: Blob;
+    name: string;
   };
 }