+++ CHANGELOG.md
@@ -7,11 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [v1.0.0-alpha.1] - 07 December 2023
+
+### What's New?
+
+-   Added Download counts for 'PRO' Users
+-   Background Blur for modal and Navigation bar
+
+### Cursors Package Update:
+
+-   Sharp angle corner & resize cursors for 'Bibata Original' (related to Bibata_Cursor#146)
+
+### Fixes
+
+-   Print username in sponsor card if name is empty
+-   Fixed Some UI bugs and Color picker modal for small devices
+
 ## [v1.0.0-alpha.0] - 23 November 2023
 
 ### What's New?
 
 -   Initial public release 🎊
 
-[unreleased]: https://github.com/ful1e5/bibata/compare/v1.0.0-alpha.0...main
+[unreleased]: https://github.com/ful1e5/bibata/compare/v1.0.0-alpha.1...main
+[v1.0.0-alpha.1]: https://github.com/ful1e5/bibata/compare/v1.0.0-alpha.0...v1.0.0-alpha.1
 [v1.0.0-alpha.0]: https://github.com/ful1e5/bibata/tree/v1.0.0-alpha.0